seperate out the filter components (OCP)
Group together where appropriate/dependent
