+++ app/assets/javascripts/rails_admin/filter-box-components/default_filter_component.js
@@ -1,0 +1,7 @@
+(function (components) {
+
+  components.default_filter_component = function (options) {
+    return {control: '<input type="text" class="input-sm form-control" name="' + options.value_name + '" value="' + options.field_value + '"/>' }
+  }
+
+})(FilterBoxComponents);
